# perf: don't try to match rules that will never match