This PR adds an optimization to the rule matching engine that will avoid evaluating rules that it can prove will never match. 

The optimization adds a pre-processing step that groups rules into "easy" and "hard" buckets. Easy rules are those that use only "easy" features, that is, features implemented via hash lookup, such as mnemonic, number, string, API, etc. Hard rules are those that use at least one "hard" feature, such as features that require a scan, like substring/regex/bytes, and rules that depend on other rules via `match` statements.

For the easy bucket, rules are indexed by the features they reference. This lets the evaluation engine quickly find the rules that overlap with a given ruleset. For example, given the features `api: send` and `api: recv`, the engine picks only rules that reference `api: send` or `api: recv` and none of the rules that deal with registry manipulation, etc. 

The rule engine evaluates the hard bucket the same way it did before. I anticipate (but did not test) that scanning features are expensive to check (for each string feature, scan with each string statement, around 577 today), so we don't try to be clever here. Also, rules that depend on other rules are tricky to handle correctly. I'd imagine there would need to be multiple rounds of matching and pruning, and therefore it would be difficult to implement and confusing to read. Therefore, we don't attempt anything special for rules with dependencies: we match them like before, evaluating each rule in topological order.

In summary, the optimization drastically cuts down on the number of rules the engine has to evaluate, leading to much higher performance:

| label                                | count(evaluations)   | avg(time)   | min(time)   | max(time)   | 
|--------------------------------------|----------------------|-------------|-------------|-------------|                                                                                                                                            
| 18c30e4 base | 108,121              | 0.28s       | 0.26s       | 0.31s       |
| 1823775 better rule selection | 39,590               | 0.14s       | 0.11s       | 0.17s       | 
| 63b47ee better rule selection + short circuiting | 24,110               | 0.08s       | 0.06s       | 0.09s       |

*(via: PMA01-01, 30 iterations)*

Adding just this optimization reduces the evaluation count by 64% and matching runtime by 50%!

But even better: this optimization is independent and orthogonal to #827/#828 (short circuiting & query optimizer) so the performance improvements are complementary. When both are enabled, evaluation count is reduced by 78% and matching runtime by 66%!

### Checklist

<!-- CHANGELOG.md has a `master (unreleased)` section. Please add bug fixes, new features, breaking changes and anything else you think is worthwhile mentioning in the release notes to this file. -->
- [ ] No CHANGELOG update needed
- [x] No new tests needed
- [x] No documentation update needed