@williballenthin 2021-11-09
I think there's probably a bug in here with `not` statements: we don't want to index those rules by features found under those statements. Will update.

Please feel free to review in the meantime.