@williballenthin 2021-11-10
There may yet be a bug here in the handling of `not:` statements: if the only statement in a rule is `not: ...`, then no features will be indexed and the rule will not be evaluated.

The easiest thing to do is to consider a `not:` statement a "hard" statement and avoid any potential mistakes here. This is reasonable because `not:` statements are rarely used (at least today).