@williballenthin 2021-11-10
Also, there are a bunch of test failures: https://github.com/mandiant/capa/runs/4159197997?check_suite_focus=true