@williballenthin 2021-11-10
I identified a number of edge cases in feature indexing having to do with `not:`, `optional:`, `0 or more:` and `count(foo): 0 or more`. In some of these cases, it isn't possible to compute the set of features that might match, so its best to consider all of these as "hard" statements and evaluate the rule conservatively.

At this point, I'm not aware of any other edge cases or concerns, so I'll re-open this PR for review and discussion for merge (assuming no test failures).