@mr-tz 2021-11-11
I ran a quick comparison across 50 files to compare the rule hits and did not notice any regressions!